+++ moorestech_client/Assets/Scripts/Client.Tests/PlaceSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6a488ea260124074adb74a601b6db89
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
