+++ moorestech_client/Assets/Scripts/Client.Tests/PlaceSystem/CommonBlockPlacePointCalculatorTest.cs
@@ -9,7 +9,7 @@ using NUnit.Framework;
 using Server.Protocol.PacketResponse;
 using UnityEngine;
 
-namespace Client.Tests
+namespace Client.Tests.PlaceSystem
 {
     public class CommonBlockPlacePointCalculatorTest
     {
