+++ moorestech_client/Assets/Scripts/Client.Tests/PlaceSystem/ConveyorOverpass.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eafbe9c71283740038e1a6cebbb724e5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
