+++ moorestech_client/Assets/Scripts/Client.Tests/PlaceSystem/ConveyorOverpass/ConveyorVerticalEnvelopeTest.cs
@@ -1,7 +1,7 @@
 using Client.Game.InGame.BlockSystem.PlaceSystem.Common.ConveyorOverpass;
 using NUnit.Framework;
 
-namespace Client.Tests
+namespace Client.Tests.PlaceSystem.ConveyorOverpass
 {
     public class ConveyorVerticalEnvelopeTest
     {
